proyecto_arqui/mqtt_suscriber/validations/requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
typing
datetime
paho-mqtt
psycopg2-binary
requests
retry